Support for different name of scripts, saving file in scripts folder

+++ scripts/uart_listener.sh
@@ -1,11 +1,12 @@
 #!/bin/bash
 set -eou pipefail
 
-saving=0
-
 # This is done to not missing tabs and spaces in scripts
-export IFS=
+export IFS=" "
+
 file="example.js"
+saving=0
+scripts_path="/home/root/trik/scripts"
 
 stop_getty() {
   # replace comment in /etc/inittab
@@ -19,32 +20,34 @@ start_getty_and_exit() {
 
 prepare_saving() {
   saving=1
-  rm $file
-  touch $file
+  if [ -e "$scripts_path/$file" ]; then
+    rm $file
+  fi
+  touch "$scripts_path/$file"
 }
 
 finish_saving() {
   saving=0
 }
 
-while read -r line < /dev/ttyS1; do 
-  echo $line
-  line_array=( $line )
-  case "${line_array[0]}" in 
-    save)
-      prepare_saving
-      ;;
-    saveend)
-      finish_saving
-      ;;
-    getty)
-      start_getty
-      ;;
-    *)
-      if [ $saving == 1 ]; then
-         echo -e "$line" >> $file
-      fi
-      ;;
-  esac
+while read -r line < /dev/ttyS1; do
+  if [ $saving == 0 ]; then
+    echo $saving
+    line_array=( $line )
+    case "${line_array[0]:-}" in 
+      save)
+        file=${line_array[1]}
+        prepare_saving
+        ;;
+      saveend)
+        finish_saving
+        ;;
+      getty)
+        start_getty
+        ;;
+    esac
+  else 
+    echo $line >> "$scripts_path/$file" 
+  fi
 done
